+++ go/internal/app/keys.go
@@ -26,6 +26,7 @@
 package app
 
 import (
+	"bytes"
 	"crypto"
 	"crypto/ed25519"
 	"crypto/rand"
@@ -34,6 +35,7 @@ import (
 	"fmt"
 	"sort"
 	"strings"
+	"time"
 
 	"github.com/urfave/cli/v2"
 	"golang.org/x/crypto/ssh"
@@ -43,7 +45,7 @@ func (a *App) keysCommand() *cli.App {
 	return &cli.App{
 		Name:            "keys",
 		Usage:           "Manage keys",
-		UsageText:       "keys <list|generate|delete|import|export>",
+		UsageText:       "keys <list|generate|delete|import|import-cert|export>",
 		Description:     "The keys command is used to manage private and public keys.",
 		HideHelpCommand: true,
 		DefaultCommand:  "list",
@@ -139,6 +141,36 @@ func (a *App) keysCommand() *cli.App {
 				},
 			},
 			{
+				Name:      "show",
+				Usage:     "Show a key",
+				UsageText: "keys show <name>",
+				Action: func(ctx *cli.Context) error {
+					if ctx.Args().Len() != 1 {
+						cli.ShowSubcommandHelp(ctx)
+						return nil
+					}
+					name := ctx.Args().Get(0)
+					key, exists := a.data.Keys[name]
+					if !exists {
+						return fmt.Errorf("unknown key %q", name)
+					}
+					pub, err := ssh.ParsePublicKey(key.Public)
+					if err != nil {
+						return err
+					}
+					a.term.Printf("Public key:\n  %s %s\nFingerprint:\n  %s\n", strings.TrimSpace(string(ssh.MarshalAuthorizedKey(pub))), name, ssh.FingerprintSHA256(pub))
+					if key.Certificate != nil {
+						cert, err := ssh.ParsePublicKey(key.Certificate)
+						if err != nil {
+							return err
+						}
+						a.term.Printf("Certificate:\n  %s %s\nDetails:\n", strings.TrimSpace(string(ssh.MarshalAuthorizedKey(cert))), name)
+						a.printCertificate(cert.(*ssh.Certificate))
+					}
+					return nil
+				},
+			},
+			{
 				Name:      "import",
 				Usage:     "Import a key",
 				UsageText: "keys import <name>",
@@ -148,8 +180,12 @@ func (a *App) keysCommand() *cli.App {
 						return nil
 					}
 					name := ctx.Args().Get(0)
-
-					files := a.importFiles("", false)
+					if _, exists := a.data.Keys[name]; exists {
+						if !a.term.Confirm(fmt.Sprintf("Key %q already exists. Overwrite?", name), false) {
+							return errors.New("aborted")
+						}
+					}
+					files := a.importFiles(".pub", false)
 					if len(files) == 0 {
 						return nil
 					}
@@ -192,6 +228,13 @@ func (a *App) keysCommand() *cli.App {
 				Name:      "export",
 				Usage:     "Export a key",
 				UsageText: "keys export <name>",
+				Flags: []cli.Flag{
+					&cli.BoolFlag{
+						Name:  "private",
+						Value: false,
+						Usage: "Export the private key.",
+					},
+				},
 				Action: func(ctx *cli.Context) error {
 					if ctx.Args().Len() != 1 {
 						cli.ShowSubcommandHelp(ctx)
@@ -202,12 +245,116 @@ func (a *App) keysCommand() *cli.App {
 					if !exists {
 						return fmt.Errorf("unknown key %q", name)
 					}
-					if !a.term.Confirm(fmt.Sprintf("You are about to export the PRIVATE key %q\nContinue?", name), false) {
-						return errors.New("aborted")
+					if ctx.Bool("private") {
+						if !a.term.Confirm(fmt.Sprintf("You are about to export the PRIVATE key %q\nContinue?", name), false) {
+							return errors.New("aborted")
+						}
+						return a.exportFile(key.Private, name+".key", "application/octet-stream")
 					}
-					return a.exportFile(key.Private, name+".key", "application/octet-stream")
+					pub, err := ssh.ParsePublicKey(key.Public)
+					if err != nil {
+						return err
+					}
+					m := strings.TrimSpace(string(ssh.MarshalAuthorizedKey(pub)))
+					out := fmt.Sprintf("%s %s\n", m, name)
+					return a.exportFile([]byte(out), name+".pub", "application/octet-stream")
+				},
+			},
+			{
+				Name:      "import-cert",
+				Usage:     "Import a certificate",
+				UsageText: "keys import-cert <key-name>",
+				Action: func(ctx *cli.Context) error {
+					if ctx.Args().Len() != 1 {
+						cli.ShowSubcommandHelp(ctx)
+						return nil
+					}
+					name := ctx.Args().Get(0)
+
+					key, exists := a.data.Keys[name]
+					if !exists {
+						return fmt.Errorf("unknown key %q", name)
+					}
+
+					files := a.importFiles(".pub", false)
+					if len(files) == 0 {
+						return nil
+					}
+					f := files[0]
+					if f.Size > 20480 {
+						return fmt.Errorf("file %q is too large: %d", f.Name, f.Size)
+					}
+					content, err := f.ReadAll()
+					if err != nil {
+						return fmt.Errorf("%q: %w", f.Name, err)
+					}
+					fmt.Fprintf(a.term, "CERT:\n%s\n", content)
+					pcert, _, _, _, err := ssh.ParseAuthorizedKey(content)
+					if err != nil {
+						return fmt.Errorf("ssh.ParsePublicKey: %v", err)
+					}
+					cert, ok := pcert.(*ssh.Certificate)
+					if !ok {
+						return fmt.Errorf("file %q does not contain a valid certificate", f.Name)
+					}
+					pub, err := ssh.ParsePublicKey(key.Public)
+					if err != nil {
+						return err
+					}
+					if !bytes.Equal(cert.Key.Marshal(), pub.Marshal()) {
+						return fmt.Errorf("the certificate in %q is for a different key", f.Name)
+					}
+					key.Certificate = cert.Marshal()
+					a.data.Keys[name] = key
+
+					if err := a.saveKeys(); err != nil {
+						return err
+					}
+					a.term.Printf("New certificate for key %q imported from %q\n", name, f.Name)
+					a.printCertificate(cert)
+					return nil
 				},
 			},
 		},
+	}
+}
+
+func (a *App) printCertificate(cert *ssh.Certificate) {
+	a.term.Printf("  Serial: %x\n", cert.Serial)
+	a.term.Printf("  Type: %s\n", cert.Type())
+	a.term.Printf("  KeyId: %s\n", cert.KeyId)
+	a.term.Printf("  ValidPrincipals: %s\n", cert.ValidPrincipals)
+	a.term.Printf("  Validity: %s - %s (UTC)\n",
+		time.Unix(int64(cert.ValidAfter), 0).UTC().Format(time.DateTime),
+		time.Unix(int64(cert.ValidBefore), 0).UTC().Format(time.DateTime))
+	if len(cert.CriticalOptions) > 0 {
+		var keys []string
+		for k := range cert.CriticalOptions {
+			keys = append(keys, k)
+		}
+		sort.Strings(keys)
+		a.term.Printf("  Critical options:\n")
+		for _, k := range keys {
+			if v := cert.CriticalOptions[k]; v != "" {
+				a.term.Printf("    %s: %s\n", k, v)
+			} else {
+				a.term.Printf("    %s\n", k)
+			}
+		}
+	}
+	if len(cert.Extensions) > 0 {
+		var keys []string
+		for k := range cert.Extensions {
+			keys = append(keys, k)
+		}
+		sort.Strings(keys)
+		a.term.Printf("  Extensions:\n")
+		for _, k := range keys {
+			if v := cert.Extensions[k]; v != "" {
+				a.term.Printf("    %s: %s\n", k, v)
+			} else {
+				a.term.Printf("    %s\n", k)
+			}
+		}
 	}
 }